Matcher: Speed up by scanning for indexes once
Instead of scanning for the next path character that matches a given query
character on each iteration of the main search loop, find all instances of each
unique query character in the path just once, on demand. Then when iterating
through query characters, all we need to do is find the next offset of a path
character past the current cursor; this can be done with a binary search over
the offsets of a particular character.