Matcher: Add C PathMatch
Adds a C extension that implements the same interface as PathMatch but with the
path matching logic done in C. The way that we find the next path character(s)
that match the query character is slightly different because it's harder to do
in C what we currently do in Ruby. I'll try addressing that in the future.

We try loading the C extension, and if it's not available then we fall back to
the Ruby implementation. For the future:

- Maybe rename the C version to just PathMatch and only load one version or the
  other (trying the C version first).
- Modify the testing setup so that we can (and do) test both the C and the Ruby
  PathMatch implementations.
- Keep experimenting with search optimizations.